Refactor caching logic in admin and authentication services; introduce CacheService for user details management
